crowdstrike/src/requirements.txt
pyobas @ git+https://github.com/OpenBAS-Platform/client-python@main
crowdstrike-falconpy==1.4.4